Created reg-uo-basic to test all flips
The new commands are not implemented yet.

I need to write them into the code. "GT" will be the smart command for that. We can define this by a room's shiftability or whatever.

gt [room] as gotorooming
gt [thing] as gotothinging
gt [number] as gotonuming

Note that the number can be + or - and - changes which direction we go

NOTE TO SELF: run i7t.pl on Ugly Oafs and get rid of the double tab at 2772/2773

Also something to parse through the table of flippies

```
while $line = <A>
{
if $line =~ /^table of flip/) { $inFlipTable = 1; next; }
if ($line !~ /[a-z]/i) { next; }
if (!$inFlipTable) { next; }
if ($inFlipTable)
print ">gt $cols[0]\ngt $cols[6] (+ or -)\n>$cols[3]\nby one point\nundo\nundo"
```

This should generate test code nicely.